pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djbabel"
authors = [
  { name="Federico Beffa", email="beffa@fbengineering.ch" },
]
description = "DJ Software playlists convertion tool"
# readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "basic-colormath",
    "mutagen",
    "pillow", # only used in ./src/djbabel/serato/overview.py
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]
keywords = ["DJ", "Serato", "Serato DJ Pro", "RekordBox", "Traktor Pro 4"]

dynamic = ["version"]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
pythonpath = "src"

[project.scripts]
djbabel = "djbabel.cli:main"

[tool.pyright]
include = ["src"]

[tool.hatch.build.targets.sdist]
include = ["/src/djbabel"]

[tool.hatch.build.targets.wheel]
packages = ["src/djbabel"]

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>[^']+)"

[project.urls]
Homepage = "https://gitlab.com/fbengineering/djbabel"
Documentation = "https://fbengineering.gitlab.io/djbabel"
